Day 9, Part 2 solution

+++ src/day09/AmplifierRunnable.java
@@ -16,7 +16,7 @@ public class AmplifierRunnable implements Runnable {
                 amplifier.processInstruction(current);
 
                 if (amplifier.hasInstructionPointerBeenModified()) {
-                    current = (int) amplifier.getInstructionPointer();
+                    current = amplifier.getInstructionPointer();
                     amplifier.resetInstructionPointer();
                 } else {
                     current += amplifier.getNumValues();
